% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measoshapes-package.R
\docType{data}
\name{measo_regions02}
\alias{measo_regions02}
\alias{measo_regions02_ll}
\alias{measo_names}
\title{Sectors and zones for MEASO.}
\description{
Data is stored in 'sf' data frame form, in a native polar projection. Longitude,latitude
version is copied with a '_ll' suffix.
}
\details{
Longer names and decomposition of each region is in 'measo_names'.
}
