% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measoshapes-package.R
\docType{package}
\name{measoshapes-package}
\alias{measoshapes}
\alias{measoshapes-package}
\title{measoshapes: Southern Ocean Shapes for 'MEASO' work.}
\description{
Provides Southern Ocean Shapes for 'MEASO' work, a Marine Ecosystem
 Assessment for the Southern Ocean.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AustralianAntarcticDivision/measoshapes}
  \item Report bugs at \url{https://github.com/AustralianAntarcticDivision/measoshapes/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{michael.sumner@aad.gov.au}

Other contributors:
\itemize{
  \item Andrew Constable [contributor]
}

}
\keyword{internal}
