% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measoshapes-package.R
\docType{package}
\name{measoshapes-package}
\alias{measoshapes}
\alias{measoshapes-package}
\title{measoshapes: Southern Ocean Shapes for 'MEASO' work.}
\description{
Provides Southern Ocean Shapes for 'MEASO' work, a Marine Ecosystem
 Assessment for the Southern Ocean.
}
\author{
\strong{Maintainer}: Michael D. Sumner \email{michael.sumner@aad.gov.au}

}
\keyword{internal}
